Extract channel opening flow
Move the state machine for the channel opening flow to a separate trait.

The event handlers are completely unchanged.
